[TrimmableTypeMap] Split NativeAOT ProGuard task
Move DGML-based NativeAOT ProGuard generation into a separate MSBuild task so GenerateProguardConfiguration remains focused on linked assembly scanning.
