+++ src/Xamarin.Android.Build.Tasks/Tasks/GenerateProguardConfiguration.cs
@@ -1,11 +1,9 @@
 #nullable enable
 
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
-using System.Xml;
 using Microsoft.Build.Framework;
 using Microsoft.Android.Build.Tasks;
 
@@ -20,10 +18,6 @@ namespace Xamarin.Android.Tasks
 		[Required]
 		public string OutputFile { get; set; } = "";
 
-		public string NativeAotDgmlFile { get; set; } = "";
-
-		public string AcwMapFile { get; set; } = "";
-
 		public override bool RunTask ()
 		{
 			var dir = Path.GetDirectoryName (OutputFile);
@@ -31,11 +25,6 @@ namespace Xamarin.Android.Tasks
 				Directory.CreateDirectory (dir);
 			}
 			using var writer = File.CreateText (OutputFile);
-
-			if (!NativeAotDgmlFile.IsNullOrEmpty ()) {
-				ProcessNativeAotDgml (writer);
-				return !Log.HasLoggedErrors;
-			}
 
 			foreach (var assembly in LinkedAssemblies) {
 				ProcessAssembly (assembly.ItemSpec, writer);
@@ -75,77 +64,6 @@ namespace Xamarin.Android.Tasks
 				// Skip non-managed assemblies
 				Log.LogDebugMessage ($"Could not read assembly '{assemblyPath}': {ex.Message}");
 			}
-		}
-
-		void ProcessNativeAotDgml (TextWriter writer)
-		{
-			if (!File.Exists (NativeAotDgmlFile)) {
-				Log.LogError ("NativeAOT DGML file '{0}' was not found.", NativeAotDgmlFile);
-				return;
-			}
-			if (AcwMapFile.IsNullOrEmpty () || !File.Exists (AcwMapFile)) {
-				Log.LogError ("ACW map file '{0}' was not found.", AcwMapFile);
-				return;
-			}
-
-			var retainedTypeKeys = LoadRetainedTypeKeysFromDgml ();
-			var javaTypes = LoadJavaTypesFromAcwMap (retainedTypeKeys);
-
-			writer.WriteLine ("# ACWs retained by NativeAOT ILC");
-			foreach (var javaTypeName in javaTypes) {
-				writer.WriteLine ($"-keep class {javaTypeName} {{ *; }}");
-			}
-
-			Log.LogMessage (MessageImportance.Low, $"Generated {javaTypes.Count} NativeAOT trimmable typemap ProGuard rules from '{NativeAotDgmlFile}'.");
-		}
-
-		SortedSet<string> LoadJavaTypesFromAcwMap (HashSet<string> retainedTypeKeys)
-		{
-			var javaTypes = new SortedSet<string> (StringComparer.Ordinal);
-			foreach (var line in File.ReadLines (AcwMapFile)) {
-				var separator = line.IndexOf (';');
-				if (separator <= 0 || separator == line.Length - 1) {
-					continue;
-				}
-				var managedTypeName = line.Substring (0, separator);
-				var javaTypeName = line.Substring (separator + 1);
-				if (retainedTypeKeys.Contains (managedTypeName)) {
-					javaTypes.Add (javaTypeName);
-				}
-			}
-			return javaTypes;
-		}
-
-		HashSet<string> LoadRetainedTypeKeysFromDgml ()
-		{
-			var typeKeys = new HashSet<string> (StringComparer.Ordinal);
-			using var reader = XmlReader.Create (NativeAotDgmlFile, new XmlReaderSettings {
-				DtdProcessing = DtdProcessing.Prohibit,
-			});
-
-			while (reader.Read ()) {
-				if (reader.NodeType != XmlNodeType.Element || reader.LocalName != "Node") {
-					continue;
-				}
-
-				var label = reader.GetAttribute ("Label");
-				if (label.IsNullOrEmpty () || !label.StartsWith ("Type metadata: [", StringComparison.Ordinal)) {
-					continue;
-				}
-
-				var assemblyStart = "Type metadata: [".Length;
-				var assemblyEnd = label.IndexOf (']', assemblyStart);
-				if (assemblyEnd < 0 || assemblyEnd == label.Length - 1) {
-					continue;
-				}
-
-				var assemblyName = label.Substring (assemblyStart, assemblyEnd - assemblyStart);
-				var managedTypeName = label.Substring (assemblyEnd + 1);
-				typeKeys.Add (managedTypeName);
-				typeKeys.Add ($"{managedTypeName}, {assemblyName}");
-			}
-
-			return typeKeys;
 		}
 
 		static bool ReferencesMonoAndroid (MetadataReader reader)
